Add unit tests for Sidebar context
