+++ client/src/components/ui/sidebar/index.ts
@@ -1,0 +1,3 @@
+export * from "./SidebarContext"
+export * from "./Sidebar"
+export * from "./SidebarMenu"
